# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/todo": {
            "get": {
                "tags": [
                    "Todo Lists"
                ],
                "summary": "Get all todo lists",
                "operationId": "f3f67a03d0db4109c215d454e8ecf5af",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Todo Lists"
                ],
                "summary": "Create a new todo list",
                "operationId": "d4c375fab9a4be2fcfea38374f9c9840",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Title of the todo list",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Todo list created successfully"
                    },
                    "400": {
                        "description": "Validation errors"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/todo/{id}": {
            "get": {
                "tags": [
                    "Todo Lists"
                ],
                "summary": "Get a todo list by ID",
                "operationId": "141e4225bedc333f0b9e3c3f40188105",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Todo list not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Todo Lists"
                ],
                "summary": "Update a todo list",
                "operationId": "65ac3e6e8c3e67e1c73c4e1e7ea52f44",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Title of the todo list",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Todo list updated successfully"
                    },
                    "404": {
                        "description": "Todo list not found"
                    },
                    "400": {
                        "description": "Validation errors"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Todo Lists"
                ],
                "summary": "Delete a todo list",
                "operationId": "9adc2c5718af5c04462b1d3fcc5a7e58",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Todo list deleted successfully"
                    },
                    "404": {
                        "description": "Todo list not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}